requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
sqlalchemy>=2.0.0
python-nmap>=0.7.1
mac-vendor-lookup>=0.1.12
websockets>=12.0
aiofiles>=23.2.1
